Replace source code toggler JavaScript with <details>
This commit replaces the `javascript:toggleSource(...)` links used to
show and hide method source with an inherently toggleable `<details>`
element.  In addition to reducing JavaScript, this fixes a small
horizontal shift that occurred when changing "show" to "hide" (and vice
versa).

The `external-link.svg` in this commit is generated from
https://iconsvg.xyz/.