RW Style unstyled code ref links
RDoc does not provide syntax to style manual code ref links with
`<code></code>`.  Wrapping code ref links in `<tt></tt>` or `++` does
not give the desired result.  For example,

  ```ruby
  # {<tt>Rails</tt>}[rdoc-ref:Rails]
  ```

is rendered as

  ```html
  {<code>Rails</code><a href="..."><code>}</code></a>
  ```

and

  ```ruby
  # <tt>{Rails}[rdoc-ref:Rails]</tt>
  ```

is rendered as

  ```html
  <code>{Rails}[rdoc-ref:Rails]</code>
  ```

This commit uses postprocessing to style unstyled code ref links that
look like code.  If the link text does not contain a space or if it
matches `/\S\(/` (e.g. `foo(bar, &block)`), it is assumed to be code,
and the text is then wrapped with `<code></code>`.

This complements @ in that it allows manual code ref links to single-word modules to be styled as if they were autolinks.