Version links to Rails guides
This uses postprocessing to version links to Rails guides.  Link URLs
are changed according to `ENV["HORO_PROJECT_VERSION"]`.  For example, if
`ENV["HORO_PROJECT_VERSION"]` is "3.2.1", then

  ```
  https://guides.rubyonrails.org/testing.html
  ```

will be changed to

  ```
  https://guides.rubyonrails.org/v3.2.1/testing.html
  ```

Or if `ENV["HORO_PROJECT_VERSION"]` is "main@1337c0d3", then the same
URL will be changed to

  ```
  https://edgeguides.rubyonrails.org/testing.html
  ```